+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - Setting feed generation mode to every episode only generates audio automatically for new articles.
 - Changing episode play button to a generate button when audio hasn't been generated yet.
 - Expandable episodes with more details and options.
+- Option to archive an episode so it doesn't appear in the podcast feed.
 
 ## V1.0.5